feat: Add isLogged function and update index and landing controllers
The `isLogged` function has been added to the `functions.php` file to check if the user is logged in. This function is then used in the `index.php` and `landing.php` controllers to determine if the user is logged in and pass the information to the views. This change improves the user experience by providing personalized content based on the user's login status.