Refactor index.php and show.php controllers, add isLogged variable
The index.php and show.php controllers have been refactored to include the $isLogged variable. This variable is used to check if the user is logged in and pass the information to the views. This change improves the code structure and ensures consistent usage of the $isLogged variable throughout the application.